Generalize context menu helpers
To prepare for building actually contextual menus.
